For omics (dr.kavousi) project.
Q1: Preprocess and dimension reduction of NAFLD RNAseq and then design a bunch of classifiers to see which one is fitter to the data
Q2: Compare the performance of machines when only RNAseq features included vs. RNAseq + Metadata features
Q3: Design 2 machines to classify a binary class (Normal / Disease) and if the output is classified as a disease, 2nd machine classifies the sub-type of the disease
Q4: Design regression for prediction of Age in the  data and then split data based on SEX to see if performance would be better or not.